ProjectOverloader: Change overloader path order
The overloader paths used to be highest priority first. Now they are
order from low to high priority.

This works well with the ConfigProvider functionality, where config of
later modules is added to existing config. This way, the last overloader
path that is added in a ConfigProvider automatically has the highest
priority.